change size_t type to int

+++ sort.h
@@ -1,9 +1,9 @@
 #ifndef SORT_H
 #define SORT_H
 
-void selection_sort(int *array, size_t array_size);
-void bubble_sort(int *array, size_t array_size);
-void insertion_sort(int *array, size_t array_size);
+void selection_sort(int *array, int array_size);
+void bubble_sort(int *array, int array_size);
+void insertion_sort(int *array, int array_size);
 void merge_sort(int *array, int l, int r);
 void quick_sort(int *array, int l, int r);
